refactor: move out function to helpers
