+++ packages/engine-ui/src/components/forms/formHelpers.js
@@ -1,3 +1,5 @@
+import moment from 'moment'
+
 const handleTextInputChange = (propertyName, callback) => (event) => {
   event.persist()
 
@@ -37,7 +39,18 @@ const handleDropdownSelect = (propertyName, callback) => ({
   }))
 }
 
+const calculateMinTime = (date, minDate) => {
+  const momentDate = moment(date || minDate)
+  const isToday = momentDate.isSame(moment(), 'day')
+  if (isToday) {
+    const nowAddOneHour = momentDate.add({ hours: 1 }).toDate()
+    return nowAddOneHour
+  }
+  return moment().startOf('day').toDate() // set to 12:00 am today
+}
+
 export default {
+  calculateMinTime,
   handleTextInputChange,
   handleContactInputChange,
   handleDropdownSelect,
